requirements.txt
argparse==1.4.0
pytest==6.2.5
pytest-mock==3.7.0
mock==4.0.3
PyYAML==6.0